pyproject.toml
[project]
name = "agents-demo"
version = "0.1.0"
description = "LaunchDarkly AI Configs tutorial with LangGraph agents"
requires-python = ">=3.11"
dependencies = [
    # Agent Framework
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "langchain-anthropic>=0.1.0",
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    # LaunchDarkly
    "launchdarkly-server-sdk>=9.0.0",
    # Vector Search
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "faiss-cpu>=1.7.0",
    # PDF Processing
    "PyMuPDF>=1.23.0",
    "tiktoken>=0.5.0",
    # UI Framework
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    # Development
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.3.31",
    # MCP Integration
    "langchain-mcp-adapters>=0.1.9",
    # Redis for caching and performance
    "redis>=5.0.0",
    "beautifulsoup4>=4.13.5",
    "mcp>=1.13.1",
    "semanticscholar>=0.10.0",
    "rank-bm25>=0.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
